% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{bootci}
\alias{bootci}
\title{Bootstrap standard error for the ICC}
\usage{
bootci(x, ...)
}
\arguments{
\item{x}{An object of class \code{iccmlm}.}

\item{...}{Further arguments passed to \code{\link[boot:boot]{boot::boot()}}.}
}
\description{
Bootstrap standard error for the ICC
}
