% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccmlm-package.r
\docType{package}
\name{iccmlm-package}
\alias{iccmlm-package}
\title{iccmlm: Calculate ICC from a multilevel model fitted by the lme4 package.}
\description{
This package provides functions to calculate the intra-class correlation
coefficient (ICC) and variance partition coefficient (VPC) from
multilevel models fitted using the lme4 package.
}
