% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{iccmodel}
\alias{iccmodel}
\title{Calculate ICC with indices (i.e. on a subset of the data)}
\usage{
iccmodel(data, indices, fit)
}
\arguments{
\item{data}{The data the model is fitted to.}

\item{indices}{The indices that boot::boot() uses to select the replicate samples.}

\item{fit}{The fitted model object which the ICC is for.}
}
\description{
Calculate ICC with indices (i.e. on a subset of the data)
}
\keyword{internal}
