% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{print.iccmlm}
\alias{print.iccmlm}
\title{print method for an object of class iccmlm}
\usage{
\method{print}{iccmlm}(x, digits = getOption("digits"),
  percent = FALSE, ...)
}
\arguments{
\item{x}{Object of class iccmlm}

\item{digits}{minimal number of \emph{significant} digits, see \code{\link{print.default}}.}

\item{percent}{Logical indicating whether the ICC should be reported as a percentage.}
}
\description{
print method for an object of class iccmlm
}
