% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{print.iccvpc}
\alias{print.iccvpc}
\title{print method for an object of class iccvpc}
\usage{
\method{print}{iccvpc}(x, digits = getOption("digits"), percent = FALSE, ...)
}
\arguments{
\item{x}{Object of class iccvpc}

\item{digits}{minimal number of \emph{significant} digits, see \code{\link{print.default}}.}

\item{percent}{Logical indicating whether the ICC should be reported as a percentage.}

\item{...}{Additional arguments passed to \code{\link{print}}.}
}
\description{
print method for an object of class iccvpc
}
\examples{
require(lme4)
fm1 <- lmer(Reaction ~ Days + (1 | Subject), sleepstudy)
summary(fm1)
print(icc(fm1, bci = TRUE, seed = 12345), percent = TRUE, digits = 2)
}
