% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{ICC from an object of class lmerMod or glmerMod}
\usage{
icc(model)
}
\arguments{
\item{model}{Object of class either lmerMod or glmerMod (fitted with family = "gaussian").}
}
\value{
Object of class iccmlm containing the ICC.
}
\description{
ICC from an object of class lmerMod or glmerMod
}
