% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Calculate ICC from an object of class lmerMod or glmerMod}
\usage{
icc(model)
}
\arguments{
\item{model}{Object of class either \code{lmerMod} or \code{glmerMod} (fitted with \code{family = "gaussian"}).}
}
\value{
Object of class \code{iccmlm} containing the ICC.
}
\description{
Calculate ICC from an object of class lmerMod or glmerMod
}
