% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{print.iccmlm}
\alias{print.iccmlm}
\title{print method for an object of class iccmlm}
\usage{
\method{print}{iccmlm}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{Object of class iccmlm}
}
\description{
print method for an object of class iccmlm
}
